Optimise size of D3D11ResourceRecord by removing rarely used pointers
* If deferred contexts map a resource they'll incur a vector resize the first
  time.

+++ renderdoc/driver/d3d11/d3d11_context.h
@@ -159,7 +159,7 @@ private:
 
   StreamReader *m_FrameReader = NULL;
 
-  map<ResourceId, int> m_MapResourceRecordAllocs;
+  map<ResourceId, size_t> m_MapResourceRecordAllocs;
 
   set<ResourceId> m_MissingTracks;
